+++ frontend/src/App.js
@@ -19,7 +19,7 @@ import ProductListScreen from './screens/ProductListScreen';
 import ProductEditScreen from './screens/ProductEditScreen';
 import OrderListScreen from './screens/OrderListScreen';
 import OrderSummaryScreen from './screens/OrderSummaryScreen';
-import OrdersScreen from './adminscreens/OrdersScreen';
+import OrdersScreen from './screens/OrdersScreen';
 import SearchScreen from './screens/SearchScreen';
 import CategoryScreen from './screens/CategoryScreen';
 import ChatbotAdmin from './screens/ChatbotAdmin';
