+++ frontend/src/index.css
@@ -116,7 +116,7 @@ h2 {
   margin: 0 auto;
 }
 body {
-  background-color: white !important;
+  background-color: #F9F9F9 !important;
 }
 .red {
   border: 1px solid red;
