+++ frontend/src/reducers/orderReducers.js
@@ -144,7 +144,7 @@ export const orderListMyReducer = (
   }
 };
 
-export const orderListReducer = (state = { orders: [] }, action) => {
+export const orderListReducer = (state = { data: [] }, action) => {
   switch (action.type) {
     case ORDER_LIST_REQUEST:
       return {
@@ -153,7 +153,7 @@ export const orderListReducer = (state = { orders: [] }, action) => {
     case ORDER_LIST_SUCCESS:
       return {
         loading: false,
-        orders: action.payload,
+        data: action.payload,
       };
     case ORDER_LIST_FAIL:
       return {
