+++ frontend/src/screens/ChatbotAdmin.js
@@ -155,348 +155,358 @@ const ChatbotAdmin = () => {
 
   return (
     <>
-      {loading ? (
-        <Loader />
-      ) : error ? (
-        <Message>{error}</Message>
-      ) : (
-        <>
-          <div>
-            <Container>
-              <h3>CHATBOT</h3>
-              <StyledRowFlex>
-                <StyledSubHeader>
-                  <StyledSubHeaderText>MESSAGE</StyledSubHeaderText>
-                </StyledSubHeader>
-                <StyledBox>
-                  <InputGroup className="m-2">
-                    <FormControl
-                      type="text"
-                      placeholder="Message"
-                      aria-describedby="basic-addon2"
-                      value={fmsg.msg}
-                      onChange={(e) => {
-                        const { value } = e.target;
-                        setFmsg((t) => {
-                          const temp = { ...t };
-                          temp.msg = value;
-                          return temp;
-                        });
-                        console.log(e.target.value);
-                      }}
-                    />
-                    <InputGroup.Append>
-                      <InputGroup.Text id="basic-addon2">
-                        <i className="fas fa-edit" />
-                      </InputGroup.Text>
-                    </InputGroup.Append>
-                  </InputGroup>
-                </StyledBox>
-              </StyledRowFlex>
-              <StyledRowFlex>
-                <StyledSubHeader>
-                  <StyledSubHeaderText>ANSWERS</StyledSubHeaderText>
-                </StyledSubHeader>
-                <StyledBox>
-                  {messages.map((msg, index) => {
-                    if (msg.length === 0) {
-                      return null;
-                    }
-                    let spacemax = msg[msg.length - 1].index;
-                    spacemax = spacemax.split(' ').length - 1;
-                    return (
-                      <Accordion
-                        defaultActiveKey="0"
-                        key={msg[0].index}
-                      >
-                        <Card style={{ backgroundColor: '#F9F9F9' }}>
-                          <Card.Header
-                            style={{
-                              backgroundColor: '#F9F9F9',
-                              padding: '0',
-                              borderBottom:
-                                '2px solid rgba(0,0,0,0.125)',
-                            }}
+      <div
+        style={{
+          minHeight: '100vh',
+          minWidth: '100vw',
+          backgroundColor: 'white',
+        }}
+      >
+        {loading ? (
+          <Loader />
+        ) : error ? (
+          <Message>{error}</Message>
+        ) : (
+          <>
+            <div>
+              <Container>
+                <h3>CHATBOT</h3>
+                <StyledRowFlex>
+                  <StyledSubHeader>
+                    <StyledSubHeaderText>MESSAGE</StyledSubHeaderText>
+                  </StyledSubHeader>
+                  <StyledBox>
+                    <InputGroup className="m-2">
+                      <FormControl
+                        type="text"
+                        placeholder="Message"
+                        aria-describedby="basic-addon2"
+                        value={fmsg.msg}
+                        onChange={(e) => {
+                          const { value } = e.target;
+                          setFmsg((t) => {
+                            const temp = { ...t };
+                            temp.msg = value;
+                            return temp;
+                          });
+                          console.log(e.target.value);
+                        }}
+                      />
+                      <InputGroup.Append>
+                        <InputGroup.Text id="basic-addon2">
+                          <i className="fas fa-edit" />
+                        </InputGroup.Text>
+                      </InputGroup.Append>
+                    </InputGroup>
+                  </StyledBox>
+                </StyledRowFlex>
+                <StyledRowFlex>
+                  <StyledSubHeader>
+                    <StyledSubHeaderText>ANSWERS</StyledSubHeaderText>
+                  </StyledSubHeader>
+                  <StyledBox>
+                    {messages.map((msg, index) => {
+                      if (msg.length === 0) {
+                        return null;
+                      }
+                      let spacemax = msg[msg.length - 1].index;
+                      spacemax = spacemax.split(' ').length - 1;
+                      return (
+                        <Accordion
+                          defaultActiveKey="0"
+                          key={msg[0].index}
+                        >
+                          <Card
+                            style={{ backgroundColor: '#F9F9F9' }}
                           >
-                            <Accordion.Toggle
-                              as={StyledBtn}
-                              variant="link"
-                              eventKey="1"
-                              onClick={() => {
-                                setArrow((a) => {
-                                  const tempa = { ...a };
-                                  if (tempa[`${index}`]) {
-                                    tempa[`${index}`] = false;
-                                  } else {
-                                    tempa[`${index}`] = true;
-                                  }
-                                  return tempa;
-                                });
-                              }}
-                            >
-                              <span style={{ fontWeight: '100' }}>
-                                {msg[0].msg}&nbsp;&nbsp;&nbsp;
-                                {arrow[`${index}`] ? (
-                                  <i
-                                    style={{ color: '#222831' }}
-                                    className="fas fa-chevron-up"
-                                  />
-                                ) : (
-                                  <i
-                                    style={{ color: '#222831' }}
-                                    className="fas fa-chevron-down"
-                                  />
-                                )}
-                              </span>
-                            </Accordion.Toggle>
-                          </Card.Header>
-                          <Accordion.Collapse eventKey="1">
-                            <Card.Body
+                            <Card.Header
                               style={{
                                 backgroundColor: '#F9F9F9',
-                                position: 'relative',
-                                paddingRight: '0px',
-                                paddingLeft: '1rem',
-                                paddingBottom: '0px',
+                                padding: '0',
+                                borderBottom:
+                                  '2px solid rgba(0,0,0,0.125)',
                               }}
                             >
-                              &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                              <span>{msg[0].info}</span>
-                              <span
+                              <Accordion.Toggle
+                                as={StyledBtn}
+                                variant="link"
+                                eventKey="1"
                                 onClick={() => {
-                                  setOpen(true);
-                                  setInd(index);
-                                  setName(msg[0].index);
-                                  setBtn(msg[0].msg);
-                                  setMsgg(msg[0].info);
-                                }}
-                                style={{
-                                  cursor: 'pointer',
-                                  position: 'absolute',
-                                  right: '10px',
-                                  color: '#929293',
+                                  setArrow((a) => {
+                                    const tempa = { ...a };
+                                    if (tempa[`${index}`]) {
+                                      tempa[`${index}`] = false;
+                                    } else {
+                                      tempa[`${index}`] = true;
+                                    }
+                                    return tempa;
+                                  });
                                 }}
                               >
-                                <i className="fas fa-edit" />
-                              </span>
-                              <span
-                                onClick={() => {
-                                  setDi(index);
-                                  setDmsg(msg[0].index);
-                                  setDelt(true);
-                                }}
+                                <span style={{ fontWeight: '100' }}>
+                                  {msg[0].msg}&nbsp;&nbsp;&nbsp;
+                                  {arrow[`${index}`] ? (
+                                    <i
+                                      style={{ color: '#222831' }}
+                                      className="fas fa-chevron-up"
+                                    />
+                                  ) : (
+                                    <i
+                                      style={{ color: '#222831' }}
+                                      className="fas fa-chevron-down"
+                                    />
+                                  )}
+                                </span>
+                              </Accordion.Toggle>
+                            </Card.Header>
+                            <Accordion.Collapse eventKey="1">
+                              <Card.Body
                                 style={{
-                                  cursor: 'pointer',
-                                  position: 'absolute',
-                                  right: '50px',
-                                  color: '#929293',
+                                  backgroundColor: '#F9F9F9',
+                                  position: 'relative',
+                                  paddingRight: '0px',
+                                  paddingLeft: '1rem',
+                                  paddingBottom: '0px',
                                 }}
                               >
-                                <i className="fas fa-trash-alt" />
-                              </span>
-                              <Additem
-                                msg={msg}
-                                setMessages={setMessages}
-                                index={index}
-                                filt={msg[0].index}
-                                messages={messages}
-                              />
-                              <br />
-                              <Acc
-                                msg={msg}
-                                spacemax={spacemax}
-                                spaces={1}
-                                filt={msg[0].index}
-                                messages={messages}
-                                setMessages={setMessages}
-                                index={index}
-                              />
-                            </Card.Body>
-                          </Accordion.Collapse>
-                        </Card>
-                      </Accordion>
-                    );
-                  })}
-                  <Additem
-                    msg={null}
-                    setMessages={setMessages}
-                    index={null}
-                    filt={null}
-                    messages={messages}
+                                &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+                                <span>{msg[0].info}</span>
+                                <span
+                                  onClick={() => {
+                                    setOpen(true);
+                                    setInd(index);
+                                    setName(msg[0].index);
+                                    setBtn(msg[0].msg);
+                                    setMsgg(msg[0].info);
+                                  }}
+                                  style={{
+                                    cursor: 'pointer',
+                                    position: 'absolute',
+                                    right: '10px',
+                                    color: '#929293',
+                                  }}
+                                >
+                                  <i className="fas fa-edit" />
+                                </span>
+                                <span
+                                  onClick={() => {
+                                    setDi(index);
+                                    setDmsg(msg[0].index);
+                                    setDelt(true);
+                                  }}
+                                  style={{
+                                    cursor: 'pointer',
+                                    position: 'absolute',
+                                    right: '50px',
+                                    color: '#929293',
+                                  }}
+                                >
+                                  <i className="fas fa-trash-alt" />
+                                </span>
+                                <Additem
+                                  msg={msg}
+                                  setMessages={setMessages}
+                                  index={index}
+                                  filt={msg[0].index}
+                                  messages={messages}
+                                />
+                                <br />
+                                <Acc
+                                  msg={msg}
+                                  spacemax={spacemax}
+                                  spaces={1}
+                                  filt={msg[0].index}
+                                  messages={messages}
+                                  setMessages={setMessages}
+                                  index={index}
+                                />
+                              </Card.Body>
+                            </Accordion.Collapse>
+                          </Card>
+                        </Accordion>
+                      );
+                    })}
+                    <Additem
+                      msg={null}
+                      setMessages={setMessages}
+                      index={null}
+                      filt={null}
+                      messages={messages}
+                    />
+                  </StyledBox>
+                </StyledRowFlex>
+                <Stylediv>
+                  <StyledButtn
+                    variant="danger"
+                    onClick={() => {
+                      setMessages(tmessages);
+                    }}
+                  >
+                    Undo Changes
+                  </StyledButtn>
+                  <StyledButtn
+                    variant="danger"
+                    onClick={() => {
+                      handleSave();
+                      setTmessages(messages);
+                    }}
+                  >
+                    Save Changes
+                  </StyledButtn>
+                </Stylediv>
+                {Loading && (
+                  <>
+                    <div style={{ marginTop: '100px' }}>
+                      <Loader />
+                    </div>
+                  </>
+                )}
+                {Error && (
+                  <>
+                    <div style={{ marginTop: '100px' }}>
+                      <Message>{Error}</Message>
+                    </div>
+                  </>
+                )}
+                {Data === 'success' && (
+                  <div
+                    style={{
+                      display: 'flex',
+                      justifyContent: 'center',
+                      alignItems: 'center',
+                      width: '100%',
+                      marginTop: '100px',
+                    }}
+                  >
+                    <span className="text-success">
+                      Your changes have been saved
+                    </span>
+                  </div>
+                )}
+              </Container>
+            </div>
+          </>
+        )}
+        {open && (
+          <>
+            <div
+              style={{
+                position: 'fixed',
+                top: '0',
+                left: '0',
+                right: '0',
+                bottom: '0',
+                backgroundColor: 'rgba(0,0,0,.7)',
+                zIndex: '1000',
+              }}
+              onClick={() => {
+                setOpen(false);
+              }}
+            />
+            <div
+              style={{
+                position: 'fixed',
+                top: '50%',
+                left: '50%',
+                transform: 'translate(-50%, -50%)',
+                backgroundColor: '#FFF',
+                padding: '50px',
+                zIndex: 1000,
+                borderRadius: '20px',
+              }}
+            >
+              <Form>
+                <Form.Group controlId="formBasicEmail">
+                  <Form.Label>Msg</Form.Label>
+                  <Form.Control
+                    type="text"
+                    placeholder="Enter Msg"
+                    name={name}
+                    value={msgg}
+                    onChange={(e) => {
+                      setMsgg(e.target.value);
+                    }}
                   />
-                </StyledBox>
-              </StyledRowFlex>
-              <Stylediv>
-                <StyledButtn
-                  variant="danger"
-                  onClick={() => {
-                    setMessages(tmessages);
-                  }}
-                >
-                  Undo Changes
-                </StyledButtn>
-                <StyledButtn
-                  variant="danger"
-                  onClick={() => {
-                    handleSave();
-                    setTmessages(messages);
-                  }}
-                >
-                  Save Changes
-                </StyledButtn>
-              </Stylediv>
-              {Loading && (
-                <>
-                  <div style={{ marginTop: '100px' }}>
-                    <Loader />
-                  </div>
-                </>
-              )}
-              {Error && (
-                <>
-                  <div style={{ marginTop: '100px' }}>
-                    <Message>{Error}</Message>
-                  </div>
-                </>
-              )}
-              {Data === 'success' && (
-                <div
-                  style={{
-                    display: 'flex',
-                    justifyContent: 'center',
-                    alignItems: 'center',
-                    width: '100%',
-                    marginTop: '100px',
-                  }}
-                >
-                  <span className="text-success">
-                    Your changes have been saved
-                  </span>
-                </div>
-              )}
-            </Container>
-          </div>
-        </>
-      )}
-      {open && (
-        <>
-          <div
-            style={{
-              position: 'fixed',
-              top: '0',
-              left: '0',
-              right: '0',
-              bottom: '0',
-              backgroundColor: 'rgba(0,0,0,.7)',
-              zIndex: '1000',
-            }}
-            onClick={() => {
-              setOpen(false);
-            }}
-          />
-          <div
-            style={{
-              position: 'fixed',
-              top: '50%',
-              left: '50%',
-              transform: 'translate(-50%, -50%)',
-              backgroundColor: '#FFF',
-              padding: '50px',
-              zIndex: 1000,
-              borderRadius: '20px',
-            }}
-          >
-            <Form>
-              <Form.Group controlId="formBasicEmail">
-                <Form.Label>Msg</Form.Label>
-                <Form.Control
-                  type="text"
-                  placeholder="Enter Msg"
-                  name={name}
-                  value={msgg}
-                  onChange={(e) => {
-                    setMsgg(e.target.value);
-                  }}
-                />
-              </Form.Group>
+                </Form.Group>
 
-              <Form.Group controlId="formBasicPassword">
-                <Form.Label>Option Name</Form.Label>
-                <Form.Control
-                  type="text"
-                  placeholder="Enter Option Name"
+                <Form.Group controlId="formBasicPassword">
+                  <Form.Label>Option Name</Form.Label>
+                  <Form.Control
+                    type="text"
+                    placeholder="Enter Option Name"
+                    name={name}
+                    value={btn}
+                    onChange={(e) => {
+                      setBtn(e.target.value);
+                    }}
+                  />
+                </Form.Group>
+                <StyledButton
+                  variant="danger"
+                  type="submit"
                   name={name}
-                  value={btn}
-                  onChange={(e) => {
-                    setBtn(e.target.value);
+                  onClick={(e) => {
+                    handleClick(e, ind);
                   }}
-                />
-              </Form.Group>
-              <StyledButton
-                variant="danger"
-                type="submit"
-                name={name}
-                onClick={(e) => {
-                  handleClick(e, ind);
-                }}
-              >
-                Done
-              </StyledButton>
-            </Form>
-          </div>
-        </>
-      )}
-      {delt && (
-        <>
-          <div
-            style={{
-              position: 'fixed',
-              top: '0',
-              left: '0',
-              right: '0',
-              bottom: '0',
-              backgroundColor: 'rgba(0,0,0,.7)',
-              zIndex: '1000',
-            }}
-            onClick={() => {
-              setDelt(false);
-            }}
-          />
-          <div
-            style={{
-              position: 'fixed',
-              top: '50%',
-              left: '50%',
-              transform: 'translate(-50%, -50%)',
-              backgroundColor: '#FFF',
-              padding: '50px',
-              zIndex: 1000,
-              borderRadius: '20px',
-            }}
-          >
-            Are you sure you want to delete this?
-            <StyledDiv>
-              <StyledButtn
-                variant="danger"
-                onClick={() => {
-                  handleDelete(dmsg, di);
-                  setDelt(false);
-                }}
-              >
-                Yes
-              </StyledButtn>
-              <StyledButtn
-                variant="danger"
-                onClick={() => {
-                  setDelt(false);
-                }}
-              >
-                No
-              </StyledButtn>
-            </StyledDiv>
-          </div>
-        </>
-      )}
+                >
+                  Done
+                </StyledButton>
+              </Form>
+            </div>
+          </>
+        )}
+        {delt && (
+          <>
+            <div
+              style={{
+                position: 'fixed',
+                top: '0',
+                left: '0',
+                right: '0',
+                bottom: '0',
+                backgroundColor: 'rgba(0,0,0,.7)',
+                zIndex: '1000',
+              }}
+              onClick={() => {
+                setDelt(false);
+              }}
+            />
+            <div
+              style={{
+                position: 'fixed',
+                top: '50%',
+                left: '50%',
+                transform: 'translate(-50%, -50%)',
+                backgroundColor: '#FFF',
+                padding: '50px',
+                zIndex: 1000,
+                borderRadius: '20px',
+              }}
+            >
+              Are you sure you want to delete this?
+              <StyledDiv>
+                <StyledButtn
+                  variant="danger"
+                  onClick={() => {
+                    handleDelete(dmsg, di);
+                    setDelt(false);
+                  }}
+                >
+                  Yes
+                </StyledButtn>
+                <StyledButtn
+                  variant="danger"
+                  onClick={() => {
+                    setDelt(false);
+                  }}
+                >
+                  No
+                </StyledButtn>
+              </StyledDiv>
+            </div>
+          </>
+        )}
+      </div>
     </>
   );
 };
